Pad formatted output when it gets shorter
When using the CurrentByte formatter the rendered progress bar may get
shorter.  When the current values advances from under 1MB to over 1MB
the bar should show "1024.00KB" then "1.00MB".

The output was not padded so the end result on the screen would look
like "1.00MB0KB" which was not easy to read and does not look nice.

Now the output is padded so the characters left over from the last
render are replaced by spaces so the result on the screen will look like
"1.00MB".

To accomplish this the existing ProgressBar#padout method is called from
ProgressBar#render and #padout is updated to use
ProgressBar#display_columns so the next render width is correctly
calculated.