Always restore hidden cursor
Through some subtleties it is possible for the progress bar to not
restore the cursor:

    b = TTY::ProgressBar.new "stuff |:bar|",
                             hide_cursor: true,
                             no_width: true,
                             total: 5,
                             width: 5

    b.finish

Prior to this commit when #finish executes it will not restore the
cursor.

Tracing through #finish, while `hide_cursor` is true, we skip restoring
the cursor because the last render width is still 0 as we have not
rendered.  We haven't completed the progress bar so we are not #done?.
We then #render.

In #render we are still not #done? so we hide the cursor because
`hide_cursor` is true, our last render width is still zero, and the
current value is not the total.

This leaves the cursor hidden when we exit the program.

This commit moves cursor restoration to the ensure block of finish so it
should always run, even if we are #done?