Fix :eta and :eta_time format tokens display when progress is not started
Logging a message with the log method throws NaN (FloatDomainError)
when progress has not started yet. This is due to elapsed time and
progress ratio both being zero before a timer is started. The division of
elapsed time by ratio returns NaN value. This fix ensures that markers
for unknown time estimation such as --s and --:--:-- are displayed instead.